Implement argument parsing on first run
Since the only argument allowed is `-h`, the code errors on
more than two args (first being the program name) or the second
arg not being `-h`.

This commit also adds the function prototype for a function that
will prove helpful for running arbitrary commands, so I don't forget
how I planned to make it happen.