Command parsing: implement `exit` command
This implementation of `exit` (partially) follows the example of
`bash`, which will exit if the command is `exit`, but will also print
an error if any invalid arguments are given to `exit`. All arguments to
`exit` are invalid in this case, so here an error is printed if anything
follows `exit`.