Implement basic output redirection
If the user's command contains a '>' followed by a single file name,
stdout is redirected to `filename.out` and stderr is redirected to
`filename.err`.

All possible errors should be accounted for in this change, though I
probably missed something.

This change also adds a `free()` for the malloc'd line from getline.